requirements.txt
torch>=1.0
argparse==1.1

numpy>=1.18.1
pkuseg>=0.0.25
sklearn>=0.0
scikit-learn>=0.22.1
pytorch-crf>=0.7.2